requirements.txt
flask>=3.0
requests>=2.31
calypsoai
sentence-transformers>=2.2
faiss-cpu>=1.7
